requirements.txt
###############################################################################
#                                 Logger
###############################################################################
structlog==17.2.0

###############################################################################
#                                  Csv
###############################################################################
xlrd==1.0.0

###############################################################################
#                                  PDFS
###############################################################################
pdfminer.six==20170419

###############################################################################
#                                  Files
###############################################################################
requests==2.9.1

###############################################################################
#                                  Dates
###############################################################################
dateparser==0.7.0
python-dateutil==2.7.2

###############################################################################
#                                   Geo
###############################################################################
six==1.10.0
pycountry==16.11.27.1
pytz==2017.2
geopy==1.11.0

###############################################################################
#                                  Strings
###############################################################################
html2text==3.200.3
beautifulsoup4==4.4.1
Markdown==2.6.2
lxml==4.0.0

###############################################################################
#                                  Tests
###############################################################################
pep8==1.5.7
py==1.4.31
pytest==3.0.2
pytest-cov==2.3.1
mock==2.0.0
pylint==1.7.2
# deps:
coverage==4.2
cov-core==1.15.0

###############################################################################
#                                Documentation
###############################################################################

alabaster==0.7.6          # via sphinx
babel==2.0                # via sphinx
docutils==0.12            # via sphinx
jinja2==2.8               # via sphinx
markupsafe==0.23          # via jinja2
pygments==2.0.2           # via sphinx
snowballstemmer==1.2.0    # via sphinx
sphinx==1.3.1             # via sphinx